List all sportsdataverse dependencies

List all sportsdataverse dependencies

Usage

sportsdataverse_deps(
  recursive = TRUE,
  pkg_list = get_core_functions(),
  repos = getOption("repos")
)

Arguments

recursive

If TRUE, will also list all dependencies of sportsdataverse packages.

pkg_list

The list of the package names to check (including the sportsdataverse package itself) Defaults to getOptions("repos").

repos

The repositories to use to check for updates. Defaults to getOptions("repos").

Examples

  try(sportsdataverse_deps() %>% knitr::kable())
#> 
#> 
#> |package      |cran    |local   |behind |
#> |:------------|:-------|:-------|:------|
#> |Rcpp         |1.0.8.3 |1.0.8.3 |FALSE  |
#> |RcppParallel |5.1.5   |5.1.5   |FALSE  |
#> |curl         |4.3.2   |4.3.2   |FALSE  |
#> |data.table   |1.14.2  |1.14.2  |FALSE  |
#> |dplyr        |1.0.9   |1.0.9   |FALSE  |
#> |furrr        |0.3.0   |0.3.0   |FALSE  |
#> |future       |1.26.1  |1.26.1  |FALSE  |
#> |ggplot2      |3.3.6   |3.3.6   |FALSE  |
#> |glue         |1.6.2   |1.6.2   |FALSE  |
#> |httr         |1.4.3   |1.4.3   |FALSE  |
#> |janitor      |2.1.0   |2.1.0   |FALSE  |
#> |jsonlite     |1.8.0   |1.8.0   |FALSE  |
#> |lubridate    |1.8.0   |1.8.0   |FALSE  |
#> |magrittr     |2.0.3   |2.0.3   |FALSE  |
#> |mgcv         |1.8.40  |1.8.40  |FALSE  |
#> |nnet         |7.3.17  |7.3.17  |FALSE  |
#> |pitchRx      |1.8.2   |1.8.2   |FALSE  |
#> |progress     |1.2.2   |1.2.2   |FALSE  |
#> |progressr    |0.10.1  |0.10.1  |FALSE  |
#> |purrr        |0.3.4   |0.3.4   |FALSE  |
#> |qdapRegex    |0.7.5   |0.7.5   |FALSE  |
#> |readr        |2.1.2   |2.1.2   |FALSE  |
#> |rlang        |1.0.3   |1.0.3   |FALSE  |
#> |rvest        |1.0.2   |1.0.2   |FALSE  |
#> |stringi      |1.7.6   |1.7.6   |FALSE  |
#> |stringr      |1.4.0   |1.4.0   |FALSE  |
#> |tibble       |3.1.7   |3.1.7   |FALSE  |
#> |tidyr        |1.2.0   |1.2.0   |FALSE  |
#> |usethis      |2.1.6   |2.1.6   |FALSE  |
#> |withr        |2.5.0   |2.5.0   |FALSE  |
#> |xml2         |1.3.3   |1.3.3   |FALSE  |